SchemaRetriever: expect that the ThingpediaClient returns ThingTalk code
The JSON formats are messy and inconsistent between schemas, metas,
manifests, etc.

Let's replace all of them by always consistently returning ThingTalk
code in the Thingpedia API, and parsing the code in SchemaRetriever.

Note that the JSON endpoints will not go away from Thingpedia though:
not all clients can consume ThingTalk. Notably, luinet is written
in Python and would struggle with ThingTalk syntax.

To simplify this commit, the test files are still in JSON. When
Thingpedia is fully converted, we can regenerate them.
