+++ lib/schema.js
@@ -13,6 +13,7 @@ const assert = require('assert');
 
 const Type = require('./type');
 const Grammar = require('./grammar_api');
+const { typeCheckClass } = require('./typecheck');
 const Ast = require('./class_def_ast');
 
 function delay(timeout) {
@@ -34,68 +35,6 @@ class DummyMemoryClient {
         this._tables.set(table, { args: args, types: types });
         return Promise.resolve();
     }
-}
-
-// Parse the semi-obsolete JSON format for schemas used
-// by Thingpedia into a FunctionDef
-function makeSchemaFunctionDef(functionType, functionName, schema, isMeta) {
-    const args = [];
-    // compat with Thingpedia API quirks
-    const types = schema.types || schema.schema;
-
-    types.forEach((type, i) => {
-        type = Type.fromString(type);
-        const argname = schema.args[i];
-        const argrequired = !!schema.required[i];
-        const arginput = !!schema.is_input[i];
-
-        let direction;
-        if (argrequired)
-            direction = Ast.ArgDirection.IN_REQ;
-        else if (arginput)
-            direction = Ast.ArgDirection.IN_OPT;
-        else
-            direction = Ast.ArgDirection.OUT;
-        const metadata = {};
-        if (isMeta) {
-            metadata.prompt = schema.questions[i] || '';
-            metadata.canonical = schema.argcanonicals[i] || argname;
-        }
-        const annotations = {};
-
-        args.push(new Ast.ArgumentDef(direction, argname,
-            type, metadata, annotations));
-    });
-
-    const metadata = {};
-    if (isMeta) {
-        metadata.canonical = schema.canonical || '';
-        metadata.confirmation = schema.confirmation || '';
-    }
-    const annotations = {};
-
-    return new Ast.FunctionDef(functionType,
-                               functionName,
-                               args,
-                               schema.is_list,
-                               schema.is_monitorable,
-                               metadata,
-                               annotations);
-}
-
-function makeSchemaClassDef(kind, schema, isMeta) {
-    const queries = {};
-    for (let name in schema.queries)
-        queries[name] = makeSchemaFunctionDef('query', name, schema.queries[name], isMeta);
-    const actions = {};
-    for (let name in schema.actions)
-        actions[name] = makeSchemaFunctionDef('action', name, schema.actions[name], isMeta);
-
-    const imports = [];
-    const metadata = {};
-    const annotations = {};
-    return new Ast.ClassDef(kind, null, queries, actions,
-                            imports, metadata, annotations);
 }
 
 module.exports = class SchemaRetriever {
@@ -186,22 +125,18 @@ module.exports = class SchemaRetriever {
             return {};
         if (!this._silent)
             console.log(`Batched ${isMeta ? 'schema-meta' : 'schema'} request for ${pending}`);
-        let resolved;
-        // FIXME clean up the TpClient API...
-        if (isMeta)
-            resolved = await this._thingpediaClient.getMetas(pending);
-        else
-            resolved = await this._thingpediaClient.getSchemas(pending);
-
+        const code = await this._thingpediaClient.getSchemas(pending, isMeta);
+        const parsed = Grammar.parse(code);
         const result = {};
-        for (var kind in resolved) {
+        await Promise.all(parsed.classes.map(async (classDef) => {
             try {
-                result[kind] = makeSchemaClassDef(kind, resolved[kind], isMeta);
-                this._classCache[isMeta].set(kind, result[kind]);
+                await typeCheckClass(classDef, this);
+                this._classCache[isMeta].set(classDef.kind, classDef);
+                result[classDef.kind] = classDef;
             } catch(e) {
-                result[kind] = e;
+                result[classDef.kind] = e;
             }
-        }
+        }));
         return result;
     }
     _ensureRequest(isMeta) {
