+++ lib/typecheck.js
@@ -607,7 +607,7 @@ function loadAllSchemas(ast, schemas, scope, classes, useMeta, isMeta = false) {
     }));
 }
 
-async function typeCheckClass(klass, schemas, isMeta = false) {
+async function typeCheckClass(klass, schemas) {
     Object.entries(klass.metadata).forEach(([name, value]) => {
         if (typeof value !== 'string')
             throw new TypeError('Invalid value type for natural language annotations');
@@ -832,7 +832,7 @@ async function typeCheckDataset(dataset, schemas, classes, getMeta = false) {
 async function typeCheckMeta(meta, schemas, getMeta = false) {
     const classes = {};
     for (let klass of meta.classes) {
-        await typeCheckClass(klass, schemas, true);
+        await typeCheckClass(klass, schemas);
         classes[klass.name] = klass;
     }
     for (let dataset of meta.datasets)
@@ -843,5 +843,6 @@ module.exports = {
     typeCheckProgram,
     typeCheckFilter,
     typeCheckPermissionRule,
-    typeCheckMeta
+    typeCheckMeta,
+    typeCheckClass
 };
